requirements.txt
absl-py==0.9.0
akshare==0.5.87
aliyun-python-sdk-core==2.13.25
aliyun-python-sdk-kms==2.11.0
appdirs==1.4.4
appnope==0.1.0
astor==0.8.1
astroid==2.4.2
attrs==19.3.0
audioread==2.1.8
autopep8==1.5.3
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
bs4==0.0.1
camelot-py==0.8.1
certifi==2020.6.20
cffi==1.14.0
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
cma==3.0.3
cmake==3.18.0
colorlog==4.2.1
crcmod==1.7
cryptography==3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
demjson==2.2.4
distlib==0.3.1
dlib==19.20.0
Elixir==0.7.1
entrypoints==0.3
et-xmlfile==1.0.1
face-alignment==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpeg-python==0.2.0
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
Flask-Babel==1.0.0
fonttools==4.13.0
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.30.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.1
identify==1.4.25
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
img2pdf==0.3.6
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonpath==0.82
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
librosa==0.7.2
llvmlite==0.31.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
moviepy==1.0.3
mplfinance==0.12.6a3
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nltk==3.5
nodeenv==1.4.0
norbert==0.2.1
notebook==6.0.3
numba==0.48.0
numpy==1.19.1
objgraph==3.4.1
opencv-python==4.3.0.36
openpyxl==3.0.4
opt-einsum==3.3.0
oss2==2.12.1
packaging==20.4
paddlehub==1.8.0
paddlepaddle==1.8.3
pandas==0.25.1
pandocfilters==1.4.2
parso==0.7.1
pathlib==1.0.1
pbr==5.4.5
pdf2image==1.13.1
pdfminer==20191125
pdfminer.six==20200720
pdfplumber==0.5.22
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
pooch==1.1.1
pre-commit==2.6.0
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
psutil==5.7.2
ptyprocess==0.6.0
pychorus==0.1
pyclipper==1.2.0
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.8
pydub==0.24.1
PyExecJS==1.5.1
pyflakes==2.2.0
pygame==1.9.6
Pygments==2.6.1
pylint==2.5.3
PyMuPDF==1.17.4
pyparsing==2.4.7
PyPDF2==1.26.0
pypinyin==0.38.1
pyrsistent==0.16.0
pytesseract==0.3.4
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
rarfile==4.0
regex==2020.7.14
requests==2.24.0
resampy==0.2.2
retrying==1.3.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.3.1
Send2Trash==1.5.0
sentencepiece==0.1.92
Shapely==1.7.0
simpleaudio==1.0.4
six==1.15.0
sortedcontainers==2.2.2
SoundFile==0.10.3.post1
soupsieve==2.0.1
spleeter==1.5.4
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
tabulate==0.8.7
Tempita==0.5.2
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.24
toml==0.10.1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
typed-ast==1.4.1
unicodecsv==0.14.1
urllib3==1.25.10
virtualenv==20.0.29
visualdl==2.0.0b8
Wand==0.6.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==0.7.2
yapf==0.26.0
zipp==3.1.0
